+++ src/ui/html.js
@@ -1,0 +1,2 @@
+// Escape per contenuti inseriti in template
+export const esc = (s) => String(s ?? '').replace(/[&<>"']/g, (c) => ({ '&': '&amp;', '<': '&lt;', '>': '&gt;', '"': '&quot;', "'": '&#39;' }[c]))